Support old `redis` configuration and new `cache_store` configuration option
