+++ lib/firebase_id_token.rb
@@ -61,7 +61,10 @@ module FirebaseIdToken
     yield configuration
     # backward compatible with the config.redis = Redis.new setup that is the old way of configuing the gem
     if configuration.redis
+      require 'firebase_id_token/certificates/redis'
       configuration.cache_store = ActiveSupport::Cache::RedisCacheStore.new(redis: configuration.redis)
+    else
+      require 'firebase_id_token/certificates/active_support'
     end
   end
